+++ journal-parsing/src/main/java/com/fnz/db2/journal/retrieve/rnrn0200/DetailedJournalReceiver.java
@@ -1,10 +1,11 @@
 package com.fnz.db2.journal.retrieve.rnrn0200;
 
+import java.math.BigInteger;
 
 public record DetailedJournalReceiver(
         JournalReceiverInfo info, 
-        Integer start, 
-        Integer end, 
+        BigInteger start, 
+        BigInteger end, 
         String nextReceiver,
         String nextDualReceiver, 
         long maxEntryLength,
